+++ release_notes.md
@@ -4,6 +4,10 @@
 3. Don't let prompt create new columns. Check that dataframe results are all in results columns.
 4. regex in commands.sql not working. SD_Prob_Take_[0-9]+ not excluded.
 5. What to do with (tournament) sections not containing user's acbl_number? concat all sections by rows? concat may be correct if other sections' results are included in matchpoint calculations.
+6. Default API should be selected from json file, if exists.
+7. Change of player number is causing error.
+8. Neil's Instant Game errored out without sidebar. Find similar.
+9. Add game percentage to about.
 
 ## Prompt Notes
 1. Does chat understand overtricks, justmade, undertricks? Or should result be used?